refactor: refactor internal types
