--- src/types/Optional.ts
@@ -1,3 +1,0 @@
-type Optional<T, K extends keyof T> = Partial<Pick<T, K>> & Omit<T, K>;
-
-export default Optional;
